Fix severe mistake that can lead to unneeded login actions on app start up
GitOrigin-RevId: f79328399ca4f2a752d52d456ce941f0cdec9dc4

+++ app/src/main/java/com/github/bumblebee202111/minusonecloudmusic/ui/common/ViewUtils.kt
@@ -7,7 +7,7 @@ import androidx.annotation.Dimension.Companion.DP
 
 object ViewUtils {
     fun dpToPx(context: Context, @Dimension(unit = DP) dp: Int): Float {
-        return dpToPx(context,dp)
+        return dpToPx(context,dp.toFloat())
     }
     fun dpToPx(context: Context, @Dimension(unit = DP) dp: Float): Float {
         val r = context.resources
